Fixed bug where session close was happening an extra time.
If a session was closed explicitly by the client, the scheduled job to close the session on expiration was not being cancelled leading to a second closing. The second closing really wasn't doing anything "bad" but it did generate extra logging that was confusing if the user had expected the session to already be closed.  Refactored some of the session "kill" operations to make them a bit more thread safe.
