+++ gremlin-server/src/test/java/org/apache/tinkerpop/gremlin/server/GremlinServerIntegrateTest.java
@@ -641,36 +641,6 @@ public class GremlinServerIntegrateTest extends AbstractGremlinServerIntegration
     }
 
     @Test
-    public void shouldHaveTheSessionTimeout() throws Exception {
-        final Cluster cluster = Cluster.build().create();
-        final Client client = cluster.connect(name.getMethodName());
-
-        final ResultSet results1 = client.submit("x = [1,2,3,4,5,6,7,8,9]");
-        final AtomicInteger counter = new AtomicInteger(0);
-        results1.stream().map(i -> i.get(Integer.class) * 2).forEach(i -> assertEquals(counter.incrementAndGet() * 2, Integer.parseInt(i.toString())));
-
-        final ResultSet results2 = client.submit("x[0]+1");
-        assertEquals(2, results2.all().get().get(0).getInt());
-
-        // session times out in 3 seconds
-        Thread.sleep(3500);
-
-        try {
-            client.submit("x[1]+2").all().get();
-            fail("Session should be dead");
-        } catch (Exception ex) {
-            final Exception cause = (Exception) ex.getCause().getCause();
-            assertTrue(cause instanceof ResponseException);
-            assertEquals(ResponseStatusCode.SERVER_ERROR_SCRIPT_EVALUATION, ((ResponseException) cause).getResponseStatusCode());
-
-            // validate that we can still send messages to the server
-            assertEquals(2, client.submit("1+1").all().join().get(0).getInt());
-        } finally {
-            cluster.close();
-        }
-    }
-
-    @Test
     public void shouldNotHavePartialContentWithOneResult() throws Exception {
         try (SimpleClient client = new WebSocketClient()) {
             final RequestMessage request = RequestMessage.build(Tokens.OPS_EVAL)
@@ -691,134 +661,6 @@ public class GremlinServerIntegrateTest extends AbstractGremlinServerIntegration
             assertEquals(1, responses.size());
         }
     }
-    
-    @Test
-    @SuppressWarnings("unchecked")
-    public void shouldExecuteInSessionAndSessionlessWithoutOpeningTransactionWithSingleClient() throws Exception {
-        assumeNeo4jIsPresent();
-        
-        try (final SimpleClient client = new WebSocketClient()) {
-
-            //open a transaction, create a vertex, commit
-            final RequestMessage openRequest = RequestMessage.build(Tokens.OPS_EVAL)
-                    .processor("session")
-                    .addArg(Tokens.ARGS_SESSION, name.getMethodName())
-                    .addArg(Tokens.ARGS_GREMLIN, "graph.tx().open()")
-                    .create();
-            final List<ResponseMessage> openResponses = client.submit(openRequest);
-            assertEquals(1, openResponses.size());
-            assertEquals(ResponseStatusCode.SUCCESS, openResponses.get(0).getStatus().getCode());
-
-            final RequestMessage addRequest = RequestMessage.build(Tokens.OPS_EVAL)
-                    .processor("session")
-                    .addArg(Tokens.ARGS_SESSION, name.getMethodName())
-                    .addArg(Tokens.ARGS_GREMLIN, "v=graph.addVertex(\"name\",\"stephen\")")
-                    .create();
-            final List<ResponseMessage> addResponses = client.submit(addRequest);
-            assertEquals(1, addResponses.size());
-            assertEquals(ResponseStatusCode.SUCCESS, addResponses.get(0).getStatus().getCode());
-
-            final RequestMessage commitRequest = RequestMessage.build(Tokens.OPS_EVAL)
-                    .processor("session")
-                    .addArg(Tokens.ARGS_SESSION, name.getMethodName())
-                    .addArg(Tokens.ARGS_GREMLIN, "graph.tx().commit()")
-                    .create();
-            final List<ResponseMessage> commitResponses = client.submit(commitRequest);
-            assertEquals(1, commitResponses.size());
-            assertEquals(ResponseStatusCode.SUCCESS, commitResponses.get(0).getStatus().getCode());
-
-            // Check to see if the transaction is closed.
-            final RequestMessage checkRequest = RequestMessage.build(Tokens.OPS_EVAL)
-                    .processor("session")
-                    .addArg(Tokens.ARGS_SESSION, name.getMethodName())
-                    .addArg(Tokens.ARGS_GREMLIN, "graph.tx().isOpen()")
-                    .create();
-            final List<ResponseMessage> checkResponses = client.submit(checkRequest);
-            assertEquals(1, checkResponses.size());
-            assertEquals(ResponseStatusCode.SUCCESS, checkResponses.get(0).getStatus().getCode());
-            assertThat(((List<Boolean>) checkResponses.get(0).getResult().getData()).get(0), is(false));
-
-            //lets run a sessionless read
-            final RequestMessage sessionlessRequest = RequestMessage.build(Tokens.OPS_EVAL)
-                    .addArg(Tokens.ARGS_GREMLIN, "graph.traversal().V()")
-                    .create();
-            final List<ResponseMessage> sessionlessResponses = client.submit(sessionlessRequest);
-            assertEquals(1, sessionlessResponses.size());
-            assertEquals(ResponseStatusCode.SUCCESS, sessionlessResponses.get(0).getStatus().getCode());
-
-            // Check to see if the transaction is still closed.
-            final RequestMessage checkAgainRequest = RequestMessage.build(Tokens.OPS_EVAL)
-                    .processor("session")
-                    .addArg(Tokens.ARGS_SESSION, name.getMethodName())
-                    .addArg(Tokens.ARGS_GREMLIN, "graph.tx().isOpen()")
-                    .create();
-            final List<ResponseMessage> checkAgainstResponses = client.submit(checkAgainRequest);
-            assertEquals(1, checkAgainstResponses.size());
-            assertEquals(ResponseStatusCode.SUCCESS, checkAgainstResponses.get(0).getStatus().getCode());
-            assertThat(((List<Boolean>) checkAgainstResponses.get(0).getResult().getData()).get(0), is(false));
-        }
-    }
-
-    @Test
-    @SuppressWarnings("unchecked")
-    public void shouldExecuteInSessionWithTransactionManagement() throws Exception {
-        assumeNeo4jIsPresent();
-
-        try (final SimpleClient client = new WebSocketClient()) {
-            final RequestMessage addRequest = RequestMessage.build(Tokens.OPS_EVAL)
-                    .processor("session")
-                    .addArg(Tokens.ARGS_SESSION, name.getMethodName())
-                    .addArg(Tokens.ARGS_GREMLIN, "v=graph.addVertex(\"name\",\"stephen\")")
-                    .addArg(Tokens.ARGS_MANAGE_TRANSACTION, true)
-                    .create();
-            final List<ResponseMessage> addResponses = client.submit(addRequest);
-            assertEquals(1, addResponses.size());
-            assertEquals(ResponseStatusCode.SUCCESS, addResponses.get(0).getStatus().getCode());
-
-            // Check to see if the transaction is closed.
-            final RequestMessage checkRequest = RequestMessage.build(Tokens.OPS_EVAL)
-                    .processor("session")
-                    .addArg(Tokens.ARGS_SESSION, name.getMethodName())
-                    .addArg(Tokens.ARGS_GREMLIN, "graph.tx().isOpen()")
-                    .create();
-            final List<ResponseMessage> checkResponses = client.submit(checkRequest);
-            assertEquals(1, checkResponses.size());
-            assertEquals(ResponseStatusCode.SUCCESS, checkResponses.get(0).getStatus().getCode());
-            assertThat(((List<Boolean>) checkResponses.get(0).getResult().getData()).get(0), is(false));
-
-            // lets run a sessionless read and validate that the transaction was managed
-            final RequestMessage sessionlessRequest = RequestMessage.build(Tokens.OPS_EVAL)
-                    .addArg(Tokens.ARGS_GREMLIN, "graph.traversal().V().values('name')")
-                    .create();
-            final List<ResponseMessage> sessionlessResponses = client.submit(sessionlessRequest);
-            assertEquals(1, sessionlessResponses.size());
-            assertEquals(ResponseStatusCode.SUCCESS, sessionlessResponses.get(0).getStatus().getCode());
-            assertEquals("stephen", ((List<String>) sessionlessResponses.get(0).getResult().getData()).get(0));
-
-            // make sure the session is intact
-            final RequestMessage getRequest = RequestMessage.build(Tokens.OPS_EVAL)
-                    .processor("session")
-                    .addArg(Tokens.ARGS_SESSION, name.getMethodName())
-                    .addArg(Tokens.ARGS_GREMLIN, "v.values(\"name\")")
-                    .addArg(Tokens.ARGS_MANAGE_TRANSACTION, true)
-                    .create();
-            final List<ResponseMessage> getResponses = client.submit(getRequest);
-            assertEquals(1, getResponses.size());
-            assertEquals(ResponseStatusCode.SUCCESS, getResponses.get(0).getStatus().getCode());
-            assertEquals("stephen", ((List<String>) getResponses.get(0).getResult().getData()).get(0));
-
-            // Check to see if the transaction is still closed.
-            final RequestMessage checkAgainRequest = RequestMessage.build(Tokens.OPS_EVAL)
-                    .processor("session")
-                    .addArg(Tokens.ARGS_SESSION, name.getMethodName())
-                    .addArg(Tokens.ARGS_GREMLIN, "graph.tx().isOpen()")
-                    .create();
-            final List<ResponseMessage> checkAgainstResponses = client.submit(checkAgainRequest);
-            assertEquals(1, checkAgainstResponses.size());
-            assertEquals(ResponseStatusCode.SUCCESS, checkAgainstResponses.get(0).getStatus().getCode());
-            assertThat(((List<Boolean>) checkAgainstResponses.get(0).getResult().getData()).get(0), is(false));
-        }
-    }
 
     @Test
     @SuppressWarnings("unchecked")
@@ -836,32 +678,5 @@ public class GremlinServerIntegrateTest extends AbstractGremlinServerIntegration
             final DetachedVertex v = ((ArrayList<DetachedVertex>) responses.get(0).getResult().getData()).get(0);
             assertEquals("jason", v.value("name"));
         }
-    }
-
-    @Test
-    public void shouldEnsureSessionBindingsAreThreadSafe() throws Exception {
-        final Cluster cluster = Cluster.open();
-        final Client client = cluster.connect(name.getMethodName());
-
-        client.submitAsync("a=100;b=1000;c=10000;null");
-        final int requests = 1000;
-        final List<CompletableFuture<ResultSet>> futures = new ArrayList<>(requests);
-        IntStream.range(0, requests).forEach(i -> {
-            try {
-                futures.add(client.submitAsync("a+b+c"));
-            } catch (Exception ex) {
-                throw new RuntimeException(ex);
-            }
-        });
-
-        assertEquals(requests, futures.size());
-
-        for(CompletableFuture<ResultSet> f : futures) {
-            final Result r = f.get().one();
-            assertEquals(11100, r.getInt());
-        }
-
-        client.close();
-        cluster.close();
     }
 }
